Moved and changed MessageConfig.java to com.example.bugtracker so it works now. html files supplemented by layout:decorate and layout:fragment.
