+++ src/main/java/com/example/bugtracker/BugTrackerApplication.java
@@ -2,12 +2,21 @@ package com.example.bugtracker;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.Bean;
+import org.springframework.web.servlet.LocaleResolver;
+import org.springframework.web.servlet.i18n.SessionLocaleResolver;
+
+import java.util.Locale;
 
 @SpringBootApplication
 public class BugTrackerApplication {
 
+
     public static void main(String[] args) {
         SpringApplication.run(BugTrackerApplication.class, args);
+
     }
+
+
 
 }
