+++ src/main/resources/application.properties
@@ -5,7 +5,6 @@ spring.datasource.driver-class-name=org.postgresql.Driver
 spring.jpa.hibernate.ddl-auto=update
 spring.jpa.show-sql=true
 
-
-spring.messages.basename=i18n/messages
+spring.messages.basename=languages/messages
 spring.messages.encoding=UTF-8
 spring.messages.fallback-to-system-locale=false
