--- src/main/resources/i18n/messages_en.properties
@@ -1,8 +1,0 @@
-# Navbar links
-navbar.home=Home1
-
-# Index links
-index.bugtracker=bt
-
-# Main Layout links
-mainLayout.bb=englishML
