--- src/main/resources/i18n/messages_pl.properties
@@ -1,8 +1,0 @@
-# Navbar links
-navbar.home=Strona2
-
-# Index links
-index.bugtracker=sledzenie robaka2
-
-# Main Layout links
-mainLayout.bb=polishML2
