+++ src/main/resources/languages/messages.properties
@@ -1,5 +1,8 @@
 # Navbar links
-navbar.home=Strona1
+navbar.home=StronaPLDefault
+navbar.lang.chooseLanguage=J?zyk
+navbar.lang.pl=polski
+navbar.lang.en=angielski
 
 # Index links
 index.bugtracker=sledzenie robaka1
