+++ src/main/resources/languages/messages_en.properties
@@ -1,0 +1,11 @@
+# Navbar links
+navbar.home=Home1
+navbar.lang.chooseLanguage=Language
+navbar.lang.pl=polish
+navbar.lang.en=english
+
+# Index links
+index.bugtracker=bugtracker
+
+# Main Layout links
+mainLayout.bb=englishML
