+++ src/main/resources/languages/messages_pl.properties
@@ -1,0 +1,11 @@
+# Navbar links
+navbar.home=StronaPL
+navbar.lang.chooseLanguage=J?zyk
+navbar.lang.pl=polski
+navbar.lang.en=angielski
+
+# Index links
+
+
+# Main Layout links
+mainLayout.bb=polskiML2
