+++ src/main/resources/templates/main-layout/navbar.html
@@ -1,36 +1,42 @@
- <div th:fragment="navbar">
+<div th:fragment="navbar">
     <nav class="navbar navbar-expand-lg navbar-light bg-light">
-    <!-- <a class="navbar-brand" href="#">Navbar</a> -->
-    <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarNavDropdown"
-            aria-controls="navbarNavDropdown" aria-expanded="false" aria-label="Toggle navigation">
-        <span class="navbar-toggler-icon"></span>
-    </button>
- <div class="collapse navbar-collapse" id="navbarNavDropdown">
-        <ul class="navbar-nav">
-            <li class="nav-item active">
-                <a class="nav-link" href="#"><b th:text="#{navbar.home}">Default Home</b></a>
-               <!-- <a class="nav-link" href="#"><b>Home</b>Home <span class="sr-only"></span></a> -->
-            </li>
-            <li class="nav-item">
-                <a class="nav-link" href="#">Features</a>
-            </li>
-            <li class="nav-item">
-                <a class="nav-link" href="#">Pricing</a>
-            </li>
-            <!--
-         <li class="nav-item dropdown">
-             <a class="nav-link dropdown-toggle" href="#" id="navbarDropdownMenuLink" role="button"
-                data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
-                 Dropdown link
-             </a>
-             <div class="dropdown-menu" aria-labelledby="navbarDropdownMenuLink">
-                 <a class="dropdown-item" href="index.html">Action</a>
-                 <a class="dropdown-item" href="#">Another action</a>
-                 <a class="dropdown-item" href="#">Something else here</a>
-             </div>
-         </li>
-         -->
-     </ul>
- </div>
-</nav>
+        <!-- <a class="navbar-brand" href="#">Navbar</a> -->
+        <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarNavDropdown"
+                aria-controls="navbarNavDropdown" aria-expanded="false" aria-label="Toggle navigation">
+            <span class="navbar-toggler-icon"></span>
+        </button>
+        <div class="collapse navbar-collapse" id="navbarNavDropdown">
+            <ul class="navbar-nav">
+                <li class="nav-item active">
+                    <a class="nav-link" href="#"><b th:text="#{navbar.home}">Default Home</b></a>
+                    <!-- <a class="nav-link" href="#"><b>Home</b>Home <span class="sr-only"></span></a> -->
+                </li>
+                <li class="nav-item">
+                    <a class="nav-link" href="#">Features</a>
+                </li>
+                <li class="nav-item">
+                    <a class="nav-link" href="#">Pricing</a>
+                </li>
+            </ul>
+
+            <div class="dropdown" mt-3>
+                <button class="btn btn-secondary dropdown-toggle" type="button" id="book-dropdown"
+                        data-bs-toggle="dropdown" th:text="#{navbar.lang.chooseLanguage}">
+                </button>
+                <ul class="dropdown-menu" aria-labelledby="book-dropdown">
+                    <li><a th:href="@{/?lang=pl}" class="dropdown-item" th:text="#{navbar.lang.pl}">Polski</a></li>
+                    <li><a th:href="@{/?lang=en}" class="dropdown-item" th:text="#{navbar.lang.en}">Angielski</a></li>
+                </ul>
+            </div>
+
+            <span th:text="#{navbar.lang.chooseLanguage}"></span>:
+            <select id="locales">
+                <option value=""></option>
+                <option value="pl" th:text="#{navbar.lang.pl}"></option>
+                <option value="en" th:text="#{navbar.lang.en}"></option>
+            </select>
+
+
+        </div>
+    </nav>
 </div>
