+++ src/main/resources/templates/projects/projectsList.html
@@ -1,10 +1,16 @@
 <!DOCTYPE html>
-<html lang="en">
+<html lang="en" layout:decorate="~{main-layout/mainLayout}"
+      xmlns:th="http://thymeleaf.org"
+      xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout">
 <head>
     <meta charset="UTF-8">
     <title>Title</title>
 </head>
 <body>
-
+<div class="container">
+    <section layout:fragment="custom-content">
+        <h1 th:text="#{index.bugtracker}">Default Message</h1>
+    </section>
+</div>
 </body>
 </html>
